requirements.txt
streamlit
pandas
google-generativeai
pillow
pdf2image